refactor(usage): extract credit-balance hook so the page clears the size gate
